activitylog/api/endpoint.go
// Copyright (c) Abstract Machines
// SPDX-License-Identifier: Apache-2.0

package api

import (
	"context"

	"github.com/absmach/magistrala/activitylog"
	"github.com/absmach/magistrala/internal/apiutil"
	"github.com/absmach/magistrala/pkg/errors"
	"github.com/go-kit/kit/endpoint"
)

func retrieveActivitiesEndpoint(svc activitylog.Service) endpoint.Endpoint {
	return func(ctx context.Context, request interface{}) (interface{}, error) {
		req := request.(retrieveActivitiesReq)
		if err := req.validate(); err != nil {
			return nil, errors.Wrap(apiutil.ErrValidation, err)
		}

		page, err := svc.RetrieveAll(ctx, req.token, req.page)
		if err != nil {
			return nil, err
		}

		return pageRes{
			ActivitiesPage: page,
		}, nil
	}
}
